Get our ApplicationController to load.
Previously, when mounting the engine in certain host apps (but
strangely, not all), our ApplicationContoller was not getting
initialized, which would cause strange issues when trying to call the
tim api in those cases.  We are supposed to get this for free, and it
always works correctly in newly-generated rails 3.2.8 apps, but older
apps that were upgraded seem to not behave correctly.  This patch
implements a simple initializer suggested on stackoverflow to make
sure the class gets loaded.

If we can later determine how to get older apps to load things
properly, we can probably drop this.