Enforce equivilence for Target Images with same Image Version
When creating a second target image for the same image
version.  We should pass in the factory base image id
to be sure that factory uses the already built base image
rather than creating a new one from the template.

This ensures that we have equivilence of Target Images
across ImageVersions.