Relax Rails version requirement.
This should make it easier for the host app to update at will, instead
of forcing them to before they are ready.  Also, general practice for
engines is to not have the lock file under version control.  To this
point we have had it to simplify development, but as it is starting to
be an issue for others, this may be a good time to remove it.